Revert "router, server 수정"
